Pass all users as props to create prog
